+++ EssentialFeed/EssentialFeediOS/Feed UI/Compossers/FeedUIComposer.swift
@@ -39,26 +39,6 @@ private extension FeedViewController {
     }
 }
 
-private final class WeakRefVirtualProxy<T: AnyObject> {
-    private weak var object : T?
-    
-    init(_ object: T) {
-        self.object = object
-    }
-}
-
-extension WeakRefVirtualProxy : FeedLoadingView where T : FeedLoadingView {
-    func display(_ viewModel: FeedLoadingViewModel) {
-        self.object?.display(viewModel)
-    }
-}
-
-extension WeakRefVirtualProxy : FeedImageView where T : FeedImageView, T.Image == UIImage {
-    func display(_ viewModel: FeedImageViewModel<UIImage>) {
-        self.object?.display(viewModel)
-    }
-}
-
 private final class FeedViewAdapter: FeedView {
     
     private weak var controller: FeedViewController?
